Anton G. Moll (Hkurn) did the yeti.
A little history: I started the yeti in the hot summer of 2009 (someone implied that I did not do it myself but had copied it (I was pissed!)), (someone thought that the initial version looked too peaceful), Bcs86 modified it to an aggressive pose, I completed the down animation & left unpixeled animation, then I left the forum and the game because the people were unfriendly and didn't contribute anything. pateame picked up and essentially finished all animations.

For some reason Skipy is mentioned here, although I think Wombat mirrored the left animation, and wrote the XML file.

Plus, I add that my work may be used only as Creative Commons BY-SA 4.0 (because GPL is for code and not art), and because of regularly changing EU copyright law a later version of CC BY-SA may be used.